Introduce a native pants client. (#11922)
* Introduce a native pants client.

Currently the client only handles talking to a pantsd brought up by
other means. The Pants repo ./pants script is updated to optionally use
the native client and prop up pantsd using the python client as needed.

Work towards #11831

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Debug macOS test failure.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Tom's feedback.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Fix parser precedence.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Build engine and client together.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Fixup flag parsing precedence / string list gathering.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Use nailgun crate to manage the pantsd ng connection.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Tighten up test CWD lock impl.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Break tests out into seperate module files.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Restructure BuildRoot to be testable.

Previously the reliance on CWD forced races across test modules which
would require a Mutex shared across those modules. That's awkward and
this arrangement provides for the important test coverage still.

[ci skip-build-wheels]

* Fix clippy lints.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Add best effort tty settings save & restore.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Use nix for termios.

[ci skip-build-wheels]

* Hack in an attempt at cross-platform process cmdline handling.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Switch from remoteprocess to sysinfo.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Use sysinfo exclusively to check pants process.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Make rust Process name extraction match Python.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Introduce ParseError to delay expensive error formatting.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Fix shell fmt and lint.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Add a client test.

Refactor launch_pantsd into a test utility to support this.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Add parse_tests.

This found a few bugs, particularly handling escape sequences for
implicit adds.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Add Args tests.

[ci skip-build-wheels]

* Fix test string formatting.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

* Add Env tests.

[ci skip-build-wheels]

* Clean up Args tests.

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]

+++ build-support/bin/rust/bootstrap_code.sh
@@ -35,16 +35,19 @@ readonly NATIVE_ENGINE_BINARY_PYO3="native_engine_pyo3.so"
 readonly NATIVE_ENGINE_RESOURCE="${REPO_ROOT}/src/python/pants/engine/internals/${NATIVE_ENGINE_BINARY}"
 readonly NATIVE_ENGINE_RESOURCE_PYO3="${REPO_ROOT}/src/python/pants/engine/internals/${NATIVE_ENGINE_BINARY_PYO3}"
 readonly NATIVE_ENGINE_RESOURCE_METADATA="${NATIVE_ENGINE_RESOURCE}.metadata"
+readonly NATIVE_CLIENT_PATH="${REPO_ROOT}/.pants"
+readonly NATIVE_CLIENT_TARGET="${NATIVE_ROOT}/target/${MODE}/pants"
 
 function _build_native_code() {
-  banner "Building native engine..."
+  banner "Building native code..."
   # NB: See Cargo.toml with regard to the `extension-module` features.
   "${REPO_ROOT}/cargo" build \
     --features=extension-module \
     --features=engine_pyo3/extension-module \
     ${MODE_FLAG} \
     -p engine \
-    -p engine_pyo3 || die
+    -p engine_pyo3 \
+    -p client || die
 }
 
 function bootstrap_native_code() {
@@ -66,7 +69,11 @@ function bootstrap_native_code() {
   if [[ -f "${NATIVE_ENGINE_RESOURCE_METADATA}" ]]; then
     engine_version_in_metadata="$(sed -n 's/^engine_version: //p' "${NATIVE_ENGINE_RESOURCE_METADATA}")"
   fi
-  if [[ ! -f "${NATIVE_ENGINE_RESOURCE}" || ! -f "${NATIVE_ENGINE_RESOURCE_PYO3}" || "${engine_version_calculated}" != "${engine_version_in_metadata}" ]]; then
+  if [[ ! -f "${NATIVE_ENGINE_RESOURCE}" || ! -f \
+    "${NATIVE_ENGINE_RESOURCE_PYO3}" || ! -f \
+    "${NATIVE_CLIENT_PATH}" || \
+    "${engine_version_calculated}" != "${engine_version_in_metadata}" ]]; then
+
     _build_native_code || die
 
     # If bootstrapping the native engine fails, don't attempt to run pants
@@ -80,15 +87,21 @@ function bootstrap_native_code() {
       die "Failed to build native engine, file missing at ${native_binary_pyo3}."
     fi
 
+    # If bootstrapping the native client fails, don't attempt to run pants afterwards.
+    if [[ ! -f "${NATIVE_CLIENT_TARGET}" ]]; then
+      die "Failed to build native client."
+    fi
+
     # Pick up Cargo.lock changes if any caused by the `cargo build`.
     engine_version_calculated="$(calculate_current_hash)"
 
     # Create the native engine resource.
     # NB: On Mac Silicon, for some reason, first removing the old native_engine.so is necessary to avoid the Pants
     #  process from being killed when recompiling.
-    rm -f "${NATIVE_ENGINE_RESOURCE}" "${NATIVE_ENGINE_RESOURCE_PYO3}"
+    rm -f "${NATIVE_ENGINE_RESOURCE}" "${NATIVE_ENGINE_RESOURCE_PYO3}" "${NATIVE_CLIENT_PATH}"
     cp "${native_binary}" "${NATIVE_ENGINE_RESOURCE}"
     cp "${native_binary_pyo3}" "${NATIVE_ENGINE_RESOURCE_PYO3}"
+    cp "${NATIVE_CLIENT_TARGET}" "${NATIVE_CLIENT_PATH}"
 
     # Create the accompanying metadata file.
     local -r metadata_file=$(mktemp -t pants.native_engine.metadata.XXXXXX)
